requirements.txt
accelerate==0.33.0
transformers==4.44.0
diffusers[torch]==0.25.0
ftfy==6.1.1
# albumentations==1.3.0
opencv-python==4.5.5.64 # 4.8.1.78
opencv-python-headless==4.5.5.64
einops==0.7.0
pytorch-lightning==1.9.0
bitsandbytes==0.44.0
lion-pytorch==0.0.6
schedulefree==1.4
pytorch-optimizer==3.5.0
prodigy-plus-schedule-free==1.9.0
prodigyopt==1.1.2
tensorboard
safetensors==0.4.4
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.24.5
# for Image utils
imagesize==1.4.1
numpy<=2.0
# for BLIP captioning
requests==2.28.2
timm==0.4.12
fairscale==0.4.4
# for WD14 captioning (tensorflow)
# tensorflow==2.10.1
# for WD14 captioning (onnx)
onnx==1.15.0
# onnxruntime-gpu==1.17.1
# onnxruntime==1.17.1
# for cuda 12.1(default 11.8)
onnxruntime-gpu --extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/

# BLIP 포함 패키지
#salesforce-lavis==1.0.2
blip==0.1.0
# Lemmatization / 어간 추출(stemming)
nltk==3.9.2

# this is for onnx: 
protobuf==3.20.3
# open clip for SDXL
open-clip-torch==2.20.0
# For logging
rich==13.7.0
# for T5XXL tokenizer (SD3/FLUX)
sentencepiece==0.2.0
# for kohya_ss library
toml==0.10.2
xformers==0.0.26.post1 --index-url https://download.pytorch.org/whl/cu121
triton==2.3.0
-e .
